#25 拆出语法树节点部分
